CCDM: remove references to removed client-api. Extract servlet container to separated article
